Version 0.0.1.228:
- Verbesserungen in VBE:
-- Anzeige in 32-bit-Modus nun funktionsfähig, kein Pixelsalat mehr, korrekte Farben. In 16/15-bit sind die Farben noch falsch.
-- "Transparentes" Bitmap nun auch transparent
-- Code vereinfacht&verbessert
- Weitere kleine Verbesserungen
- Weitere Testergebnisse

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@812 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/filesystem/fsmanager.c
@@ -47,9 +47,9 @@ file_t* fopen(const char* path, const char* mode)
         free(file);
         return(0);
     }
-    file->EOF    = false;
-    file->error  = CE_GOOD;
-    file->name   = malloc(strlen(getFilename(path))+1, 0, "fsmgr-filename");
+    file->EOF   = false;
+    file->error = CE_GOOD;
+    file->name  = malloc(strlen(getFilename(path))+1, 0, "fsmgr-filename");
     strcpy(file->name, getFilename(path));
 
     bool appendMode = false; // Used to seek to end
@@ -170,7 +170,7 @@ size_t fwrite(const void* src, size_t size, size_t count, file_t* file)
     size_t i = 0;
     for(; i < count; i++)
     {
-        for(int j = 0; j < size; j++)
+        for(size_t j = 0; j < size; j++)
         {
             fputc(((uint8_t*)src)[i*size+j], file);
         }
